[EDELIVERY-6596] added dedicated Validator class for Payload profile

+++ Domibus-MSH/src/main/java/eu/domibus/common/model/configuration/PayloadProfile.java
@@ -88,8 +88,8 @@ public class PayloadProfile extends AbstractBaseEntity {
      * @param value allowed object is
      *              {@link BigInteger }
      */
-    public void setMaxSize(final int value) {
-        this.maxSize = value;
+    public void setMaxSize(final long value) {
+        this.maxSize = Math.toIntExact(value);
     }
 
     void init(final Configuration configuration) {
